.vimrc: Delete plugins I don't use, and add some minor changes
First of all, disable S-L in normal mode, because it confuses me.
And set the colorscheme to vscode, because it has more contrast. Change
the color of the "--INSERT--" indicator, because we use block cursors
for both insert mode and normal mode.

About the removal of plugins, I removed bookmark plugin because nobody
uses it anymore, everything is scrambled after a git pull. I disabled
C-W in insert mode because I need to use this prefix to switch windows,
and if I don't, waiting for too long before hitting the next key after
C-W will just delete a previous word.

"=== STABLE END ===" thing is stupid so I deleted it.

Also delete LuaSnip plugin (haven't used it in 6 months).

Signed-off-by: Howard Chu <howardchu95@gmail.com>